+++ integrations/sonarqube/CHANGELOG.md
@@ -11,4 +11,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Features
 
-- Implemented Sonarqube integration using Ocean (#0)
+- Implemented Sonarqube integration using Ocean